+++ pdf/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.4
 * Add available dimensions for PdfPageFormat
+* Add Document properties
 
 # 1.3.3
 * Fix a bug with the RichText Widget