+++ printing/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.0
 * Breaking change: Switch libraries to AndroidX
+* Add Page information to PdfDoc object
 
 # 1.3.5
 * Restore compatibility with Flutter 1.0.0